Create ErrorBoundary Container
